Move elements-related code to separate file
